+++ test/parsimmon.test.js
@@ -153,6 +153,71 @@ suite('parser', function() {
     });
   });
 
+  suite('Parsimmon.createLanguage', function() {
+    test('should return an object of parsers', function() {
+      var lang = Parsimmon.createLanguage({
+        a: function() {
+          return Parsimmon.string('a');
+        },
+        b: function() {
+          return Parsimmon.string('b');
+        }
+      });
+      assert.ok(Parsimmon.isParser(lang.a));
+      assert.ok(Parsimmon.isParser(lang.b));
+    });
+    test('should allow direct recursion in parsers', function() {
+      var lang = Parsimmon.createLanguage({
+        Parentheses: function(r) {
+          return Parsimmon.alt(
+            Parsimmon.string('()'),
+            Parsimmon.string('(')
+              .then(r.Parentheses)
+              .skip(Parsimmon.string(')'))
+          );
+        }
+      });
+      lang.Parentheses.tryParse('(((())))');
+    });
+    test('should allow indirect recursion in parsers', function() {
+      var lang = Parsimmon.createLanguage({
+        Value: function(r) {
+          return Parsimmon.alt(
+            r.Number,
+            r.Symbol,
+            r.List
+          );
+        },
+        Number: function() {
+          return Parsimmon.regexp(/[0-9]+/).map(Number);
+        },
+        Symbol: function() {
+          return Parsimmon.regexp(/[a-z]+/);
+        },
+        List: function(r) {
+          return Parsimmon.string('(')
+            .then(Parsimmon.sepBy(r.Value, r._))
+            .skip(Parsimmon.string(')'));
+        },
+        _: function() {
+          return Parsimmon.optWhitespace;
+        }
+      });
+      lang.Value.tryParse('(list 1 2 foo (list nice 3 56 989 asdasdas))');
+    });
+  });
+
+  suite('parser.thru', function() {
+    test('should return wrapper(this)', function() {
+      function arrayify(x) {
+        return [x];
+      }
+      var parser = Parsimmon.string('');
+      var array = parser.thru(arrayify);
+      assert.strictEqual(array[0], parser);
+    });
+  });
+
   suite('Parsimmon.lookahead', function() {
     test('should handle a string', function() {
       lookahead('');
@@ -1084,6 +1149,38 @@ suite('parser', function() {
         end: {offset: 3, line: 2, column: 3}
       }
       );
+  });
+
+  test('.node(name)', function() {
+    var ys = regex(/^y*/).node('Y');
+    var parser = optWhitespace.then(ys).skip(optWhitespace);
+    assert.deepEqual(
+      parser.parse('').value,
+      {
+        name: 'Y',
+        value: '',
+        start: {offset: 0, line: 1, column: 1},
+        end: {offset: 0, line: 1, column: 1}
+      }
+    );
+    assert.deepEqual(
+      parser.parse(' yy ').value,
+      {
+        name: 'Y',
+        value: 'yy',
+        start: {offset: 1, line: 1, column: 2},
+        end: {offset: 3, line: 1, column: 4}
+      }
+    );
+    assert.deepEqual(
+      parser.parse('\nyy ').value,
+      {
+        name: 'Y',
+        value: 'yy',
+        start: {offset: 1, line: 2, column: 1},
+        end: {offset: 3, line: 2, column: 3}
+      }
+    );
   });
 
   suite('smart error messages', function() {